android: binder: add extra size to allocator.
The binder_buffer allocator currently only allocates
space for the data and offsets buffers of a Parcel.
This change allows for requesting an additional chunk
of data in the buffer, which can for example be used
to hold additional meta-data about the transaction
(eg a security context).

Change-Id: I58ab9c383a2e1a3057aae6adaa596ce867f1b157
Signed-off-by: Martijn Coenen <maco@google.com>
Signed-off-by: Marco Zanin (B--B) <mrczn.bb@gmail.com>
(cherry picked from commit 68646be576643b6ecc49d1bad441fd144e2e1637)